Customers Address Implemented

+++ src/app/components/cart-details/cart-details.component.ts
@@ -1,14 +1,15 @@
 import { Component, OnInit } from '@angular/core';
-import { MatDialog } from '@angular/material/dialog';
+
 import { MatIconRegistry } from '@angular/material/icon';
 import { DomSanitizer } from '@angular/platform-browser';
-import { Router } from '@angular/router';
-import { Observable, map } from 'rxjs';
-import { BookService } from 'src/app/services/bookservice/book.service';
-import { DataService } from 'src/app/services/dataservice/data.service';
-import { BookObj } from 'src/assets/bookInterface';
 import { DROP_DOWN, LOCATION_ICON } from 'src/assets/svg-icons';
 import { LoginSignupComponent } from '../login-signup/login-signup.component';
+import { Router } from '@angular/router';
+import { MatDialog } from '@angular/material/dialog';
+import { BookObj } from 'src/assets/bookInterface';
+import { DataService } from 'src/app/services/dataservice/data.service';
+import { BookService } from 'src/app/services/bookservice/book.service';
+import { Subscription } from 'rxjs';
 
 @Component({
   selector: 'app-cart-details',
@@ -18,9 +19,14 @@ import { LoginSignupComponent } from '../login-signup/login-signup.component';
 export class CartDetailsComponent implements OnInit {
   showAddressDetails: boolean = false;
   showOrderSummary: boolean = false;
-  cartItems$!: Observable<(BookObj & { quantity: number })[]>;
-  count: number = 1;
+  showOrderPlaced: boolean = false;
   cartItems: (BookObj & { quantity: number })[] = [];
+  authToken: string | null = null;
+  count: number = 1;
+  placeOrder: boolean = false;
+  address: any[] = [];
+  selectedAddressId: number | null = null;
+  private orderSummarySubscription!: Subscription;
 
   constructor(
     private dataService: DataService,
@@ -40,6 +46,15 @@ export class CartDetailsComponent implements OnInit {
     );
   }
 
+  ngOnInit(): void {
+    this.dataService.currCartList.subscribe((response) => {
+      this.cartItems = response;
+    });
+    this.orderSummarySubscription =
+      this.dataService.orderSummaryToggled.subscribe(() => {
+        this.toggleOrderSummary();
+      });
+  }
   toggleAddressDetails() {
     this.showAddressDetails = !this.showAddressDetails;
   }
@@ -47,39 +62,16 @@ export class CartDetailsComponent implements OnInit {
   toggleOrderSummary() {
     this.showOrderSummary = !this.showOrderSummary;
   }
-
-  ngOnInit(): void {
-    const authToken = localStorage.getItem('authToken');
-    if (authToken) {
-      // Auth token is present, fetch all cart details
-      this.bookService
-        .getAllCartDetails()
-        .pipe(map((response) => response.data))
-        .subscribe((cartItems: (BookObj & { quantity: number })[]) => {
-          this.dataService.setCartItems(cartItems); // Directly set cart items
-        });
-    }
-    this.cartItems$ = this.dataService.getCartItems();
-  }
-
-  increaseCount(book: BookObj) {
-    this.bookService
-      .updateBookQuantity(book, (book.quantity ?? 0) + 1)
-      .subscribe();
-    this.dataService.addToCart(book, 1);
-  }
-
-  decreaseCount(book: BookObj) {
-    if (book && book.quantity !== undefined && book.quantity > 1) {
-      this.bookService
-        .updateBookQuantity(book, (book.quantity ?? 0) - 1)
-        .subscribe();
-      this.dataService.addToCart(book, -1);
-    }
-  }
   removeFromCart(book: BookObj) {
+    this.cartItems = this.cartItems.filter(
+      (item) => item.bookId !== book.bookId
+    );
     this.dataService.removeFromCart(book);
-    if (book && book.bookId !== undefined) {
+    if (
+      book &&
+      book.bookId !== undefined &&
+      localStorage.getItem('authToken')
+    ) {
       this.bookService.deleteBookFromCart(book.bookId).subscribe(
         () => {},
         (error) => {
@@ -88,12 +80,43 @@ export class CartDetailsComponent implements OnInit {
       );
     }
   }
-  handlePlaceOrder(data: any, choice?: string) {
+
+  increaseCount(book: BookObj) {
+    if (book && book.quantity !== undefined) {
+      if (localStorage.getItem('authToken') != null) {
+        this.bookService.updateBookQuantity(book, ++book.quantity).subscribe(
+          () => {},
+          (error) => {
+            console.error('Error updating quantity:', error);
+          }
+        );
+      } else {
+        this.dataService.addToCart(book, 1);
+      }
+    }
+  }
+
+  decreaseCount(book: BookObj) {
+    if (book && book.quantity !== undefined && book.quantity > 1) {
+      if (localStorage.getItem('authToken') != null) {
+        this.bookService.updateBookQuantity(book, --book.quantity).subscribe(
+          () => {},
+          (error) => {
+            console.error('Error updating quantity:', error);
+          }
+        );
+      } else {
+        this.dataService.addToCart(book, -1);
+      }
+    }
+  }
+
+  handlePlaceOrder() {
     if (localStorage.getItem('authToken') != null) {
+      this.showAddressDetails = !this.showAddressDetails;
+      this.placeOrder = !this.placeOrder;
     } else {
-      // Navigate to login/signup page
       this.router.navigate(['/books']).then(() => {
-        // Open dialog after navigation
         const dialogRef = this.dialog.open(LoginSignupComponent, {
           data: { value: 'placeOrder', cart: this.cartItems },
         });
@@ -102,5 +125,8 @@ export class CartDetailsComponent implements OnInit {
         });
       });
     }
+  }
+  handleCheckout() {
+    this.router.navigate(['/orders']);
   }
 }
